feat - add USB client
Remove task for old USB command

PS : This commit doesn't compile !
