+++ src/bin/gamepad.rs
@@ -4,13 +4,12 @@ use software::libs::data::DataStore;
 use software::libs::pipeline::{run_pipeline, Pipeline};
 
 use software::libs::tasks::inputs::gamepad::GamepadInputTask;
-use software::libs::tasks::outputs::sim_commands::SimCommandsOutputTask;
-use software::libs::tasks::outputs::usb_commands::UsbCommandsOutputTask;
 use software::libs::tasks::task::Task;
 
 #[macro_use]
 extern crate log;
 use env_logger::Env;
+use software::inputs_outputs::output::OutputTask;
 
 fn main() {
     let env = Env::default()
@@ -25,11 +24,7 @@ fn main() {
 
     let mut pipeline: Pipeline<dyn Task> = vec![
         GamepadInputTask::with_cli_boxed(&mut cli),
-        if cli.real {
-            UsbCommandsOutputTask::with_cli_boxed(&mut cli)
-        } else {
-            SimCommandsOutputTask::with_cli_boxed(&mut cli)
-        },
+        OutputTask::with_cli_boxed(&mut cli),
     ];
 
     run_pipeline(&mut data_store, &mut pipeline);
