+++ src/inputs_outputs/ubs_client.rs
@@ -1,12 +1,108 @@
 use crate::inputs_outputs::output::OutputCommandSending;
+use crate::libs::cli::Cli;
 use crate::libs::constants::NUMBER_OF_ROBOTS;
-use crate::libs::data::Command;
+use crate::libs::data::{Command, KICK};
+use crate::libs::protobuf::robot_packet::IaToMainBoard;
 use crate::libs::robot::AllyRobotInfo;
+use clap::command;
+use log::{debug, error, info};
+use prost::Message;
+use serialport::SerialPort;
+use std::time::Duration;
 
-pub struct USBClient;
+pub struct USBClientCli {
+    /// USB port of the mainboard
+    #[arg(long, default_value = "/dev/USB0")]
+    usb_port: String,
+}
+
+pub struct USBClient {
+    port: Box<dyn SerialPort>,
+}
+
+impl USBClient {
+    fn prepare_packet(&mut self, command: &mut Command) -> IaToMainBoard {
+        let (kicker_cmd, kick_power) = match command.kick {
+            None => {
+                (0, 0.0 as f32) // TODO : Remove this 0 and use the kicker enum
+            }
+            Some(KICK::StraightKick { power }) => (1, power),
+            Some(KICK::ChipKick { power }) => (2, power),
+        };
+
+        let packet = IaToMainBoard {
+            robot_id: command.id,
+            normal_speed: command.forward_velocity,
+            tangential_speed: command.left_velocity,
+            angular_speed: command.angular_velocity,
+            motor_break: false,
+            kicker_cmd,
+            kick_power,
+            charge: command.charge,
+            dribbler: command.dribbler > 0.0,
+        };
+
+        let mut buf = Vec::new();
+        buf.reserve(packet.encoded_len() + 1);
+        buf.push(packet.encoded_len() as u8);
+        packet.encode(&mut buf).unwrap();
+
+        info!("{}", packet.encoded_len());
+    }
+
+    fn send(&mut self, packet: IaToMainBoard) {
+        match self.port.write(&buf[0..packet.encoded_len() + 1]) {
+            Ok(_v) => {
+                debug!("sent order: {:?}", packet);
+            }
+            Err(e) => {
+                error!("{}", e);
+            }
+        }
+    }
+
+    // TODO : Add receive function
+    fn receive(&self) -> Vec<AllyRobotInfo> {
+        vec![]
+    }
+}
 
 impl OutputCommandSending for USBClient {
-    fn send(commands: [Command; NUMBER_OF_ROBOTS]) -> AllyRobotInfo {
-        todo!()
+    fn with_cli(cli: &mut Cli) -> Self {
+        Self {
+            port: serialport::new(cli.usb_commands.usb_port.clone(), 115_200)
+                .timeout(Duration::from_millis(1))
+                .open()
+                .expect("Failed to open port"),
+        }
+    }
+
+    fn step(&mut self, mut commands: Vec<Command>) -> Vec<AllyRobotInfo> {
+        for command in commands.iter_mut() {
+            let packet = self.prepare_packet(command);
+            self.send(packet);
+            self.receive()
+        }
+        vec![]
+    }
+}
+
+impl Drop for USBClient {
+    fn drop(&mut self) {
+        let mut commands: Vec<Command> = vec![];
+
+        for i in 0..NUMBER_OF_ROBOTS {
+            commands.push(Command {
+                id: i as u32,
+                forward_velocity: 0.0,
+                left_velocity: 0.0,
+                angular_velocity: 0.0,
+                charge: false,
+                kick: None,
+                dribbler: 0.0,
+            })
+        }
+
+        self.step(commands);
     }
 }
