+++ src/libs/tasks/outputs/mod.rs
@@ -1,3 +1,1 @@
-pub mod sim_commands;
-pub mod usb_commands;
 pub mod zmq;
